+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/AuthhelperParam.java
@@ -38,12 +38,13 @@ public class AuthhelperParam extends AbstractParam {
     private static final String DEMO_MODE_KEY = AUTO_KEY + ".demo";
     private static final String RECORD_DIAGNOSTICS_KEY = AUTO_KEY + ".diagnostics";
     private static final String STEP_KEY = AUTO_KEY + ".steps.step";
+    private static final String STEP_DELAY_KEY = AUTO_KEY + ".stepDelay";
 
     private String loginUrl;
     private String username;
     private String browser;
     private int wait = DEFAULT_WAIT;
-    private boolean demoMode;
+    private int stepDelay;
     private boolean recordDiagnostics;
     private List<AuthenticationStep> steps = List.of();
 
@@ -55,7 +56,7 @@ public class AuthhelperParam extends AbstractParam {
         this.username = this.getString(USERNAME_KEY, null);
         this.browser = this.getString(BROWSER_KEY, DEFAULT_BROWSER.getId());
         this.wait = getInteger(WAIT_KEY, DEFAULT_WAIT);
-        this.demoMode = getBoolean(DEMO_MODE_KEY, false);
+        this.stepDelay = getInteger(STEP_DELAY_KEY, 0);
         this.recordDiagnostics = getBoolean(RECORD_DIAGNOSTICS_KEY, false);
 
         steps =
@@ -64,6 +65,10 @@ public class AuthhelperParam extends AbstractParam {
                         .map(AuthenticationStep::decode)
                         .filter(Objects::nonNull)
                         .toList();
+        if (this.getConfig().containsKey(DEMO_MODE_KEY)) {
+            // No longer used
+            this.getConfig().clearProperty(DEMO_MODE_KEY);
+        }
     }
 
     public String getLoginUrl() {
@@ -73,15 +78,6 @@ public class AuthhelperParam extends AbstractParam {
     public void setLoginUrl(String loginUrl) {
         this.loginUrl = loginUrl;
         getConfig().setProperty(LOGIN_URL_KEY, loginUrl);
-    }
-
-    public boolean isDemoMode() {
-        return demoMode;
-    }
-
-    public void setDemoMode(boolean demoMode) {
-        this.demoMode = demoMode;
-        getConfig().setProperty(DEMO_MODE_KEY, demoMode);
     }
 
     public String getBrowser() {
@@ -109,6 +105,15 @@ public class AuthhelperParam extends AbstractParam {
     public void setWait(int wait) {
         this.wait = wait;
         getConfig().setProperty(WAIT_KEY, wait);
+    }
+
+    public int getStepDelay() {
+        return stepDelay;
+    }
+
+    public void setStepDelay(int stepDelay) {
+        this.stepDelay = stepDelay;
+        getConfig().setProperty(STEP_DELAY_KEY, stepDelay);
     }
 
     public boolean isRecordDiagnostics() {